+++ packages/web3-eth-abi/src/coders/encode.ts
@@ -17,7 +17,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import { AbiError } from 'web3-errors';
 import { AbiInput } from 'web3-types';
 import { utils } from 'web3-validator';
-import { encodeTuple } from './base';
+import { encodeTuple } from './base/index.js';
 import { toAbiParams } from './utils.js';
 
 export function encodeParameters(abi: ReadonlyArray<AbiInput>, params: unknown[]): string {