+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -142,3 +142,7 @@ Documentation:
 ### Added
 
 -   Added public function `privateKeyToPublicKey`
+
+### Fixed
+
+-   Fixed `recover` function, `v` will be normalized to value 0,1 (#6344)