+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -138,3 +138,11 @@ Documentation:
 
 
 ## [Unreleased]
+
+### Added
+
+-   Added public function `privateKeyToPublicKey`
+
+### Fixed
+
+-   Fixed `recover` function, `v` will be normalized to value 0,1 (#6344)