+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -137,7 +137,7 @@ Documentation:
 -   Fixed "The `r` and `s` returned by `sign` to does not always consist of 64 characters" (#6411)
 
 
-## [Unreleased]
+## [4.1.0]
 
 ### Added
 
@@ -148,3 +148,5 @@ Documentation:
 ### Fixed
 
 -   Fixed `recover` function, `v` will be normalized to value 0,1 (#6344) 
+
+## [Unreleased]